chefoeuvre/requirements.txt
django==4.1.7
pytest-django==4.5.2